Make null-branch check version-strict; split test base helpers
Replace the loose Is3_0||Is3_1 null-branch predicate with a single
instance dispatcher that strictly picks the marker shape for the
document's declared OpenAPI version, so WalkSchemaLayers no longer
silently skips a 3.1 {"type":"null"} branch in a 3.0 document (or
vice versa). Rename the per-version helpers to IsNullBranch3_0 /
IsNullBranch3_1 and document the trio in XML.

Move pure JSON helpers out of OpenApiDocumentTestsBase into three
focused static classes — SchemaNavigation, SchemaValueReader,
ComponentSchemas — pulled in via using static. Trims the base from
~1213 to ~1105 lines without changing call sites.
